requirements.txt
alembic==1.1.0
astroid==2.2.5
Click==7.0
colorama==0.4.1
Flask==1.1.1
Flask-Migrate==2.5.2
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.0
gunicorn==19.9.0
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.1
lazy-object-proxy==1.4.2
Mako==1.1.0
MarkupSafe==1.1.1
mccabe==0.6.1
psycopg2==2.8.3
pylint==2.3.1
python-dateutil==2.8.0
python-editor==1.0.4
redis==3.3.8
rq==1.1.0
six==1.12.0
SQLAlchemy==1.3.8
typed-ast==1.4.0
Werkzeug==0.15.6
wrapt==1.11.2
